Rename EditCommandTestUtil to CommandTestUtil
The constants declared in EditCommandTestUtil can be used by other
command tests.

Let's rename EditCommandTestUtil to a more generic name,
CommandTestUtil.
