+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -8,15 +8,15 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_ADDRESS_AMY;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_EMAIL_AMY;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_NAME_AMY;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_PHONE_AMY;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_TAG_FRIEND;
-import static seedu.address.testutil.EditCommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.testutil.CommandTestUtil.VALID_ADDRESS_AMY;
+import static seedu.address.testutil.CommandTestUtil.VALID_ADDRESS_BOB;
+import static seedu.address.testutil.CommandTestUtil.VALID_EMAIL_AMY;
+import static seedu.address.testutil.CommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.address.testutil.CommandTestUtil.VALID_NAME_AMY;
+import static seedu.address.testutil.CommandTestUtil.VALID_PHONE_AMY;
+import static seedu.address.testutil.CommandTestUtil.VALID_PHONE_BOB;
+import static seedu.address.testutil.CommandTestUtil.VALID_TAG_FRIEND;
+import static seedu.address.testutil.CommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.testutil.TypicalPersons.INDEX_FIRST_PERSON;
 import static seedu.address.testutil.TypicalPersons.INDEX_SECOND_PERSON;
 import static seedu.address.testutil.TypicalPersons.INDEX_THIRD_PERSON;
